Offers and initial README and `Brewfile` (#7)
TL;DR
-----

Helps others use this repository with some docs and ergnomics

Details
-------

Starts to document the demonstration with a short README that will help
someone familiar with Replicated and the available tooling use it.
There's enough information for an experienced person to use what they
know to build and release the application, but not enough to get brand
new to Replicated going.

Also simplifies getting started by providing a `Brewfile` to assure the
required CLIs are installed.

To Do
-----

Expand the README to be usable by someone who's less familiar with the
Replicated platform.